refactored tests dir
